Refactor: 메인 홈 > 보드게임 도메인으로 패키지 이동
