+++ src/main/java/com/boardgo/domain/boardgame/controller/HomeBoardGameController.java
@@ -1,10 +1,10 @@
-package com.boardgo.domain.home.controller;
+package com.boardgo.domain.boardgame.controller;
 
 import static com.boardgo.common.constant.HeaderConstant.API_VERSION_HEADER1;
 
 import com.boardgo.domain.boardgame.entity.SituationType;
-import com.boardgo.domain.home.controller.response.SituationBoardGameResponse;
-import com.boardgo.domain.home.service.HomeQueryUseCase;
+import com.boardgo.domain.boardgame.service.HomeBoardGameQueryUseCase;
+import com.boardgo.domain.boardgame.service.response.SituationBoardGameResponse;
 import jakarta.validation.constraints.NotNull;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -17,17 +17,17 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/home")
-public class HomeController {
+public class HomeBoardGameController {
 
-    private final HomeQueryUseCase homeQueryUseCase;
+    private final HomeBoardGameQueryUseCase homeBoardGameQueryUseCase;
 
     @GetMapping(value = "/situation", headers = API_VERSION_HEADER1)
     public ResponseEntity<List<SituationBoardGameResponse>> getSituationBoardGame(
             @RequestParam("situationType") @NotNull SituationType situationType) {
-        return ResponseEntity.ok().body(homeQueryUseCase.getSituationBoardGame(situationType));
+        return ResponseEntity.ok()
+                .body(homeBoardGameQueryUseCase.getSituationBoardGame(situationType));
     }
 
-    // FIXME: 보드게임 도메인으로 이동
     //    @GetMapping(value = "/cumulative-popularity", headers = API_VERSION_HEADER1)
     //    public ResponseEntity<List<CumulativePopularityResponse>> getCumulativePopularity() {
     //        return ResponseEntity.ok().body(homeQueryUseCase.getCumulativePopularity());
