+++ src/main/java/com/boardgo/domain/boardgame/repository/BoardGameDslRepository.java
@@ -1,7 +1,7 @@
 package com.boardgo.domain.boardgame.repository;
 
 import com.boardgo.domain.boardgame.controller.request.BoardGameSearchRequest;
-import com.boardgo.domain.boardgame.repository.dto.SituationBoardGameDto;
+import com.boardgo.domain.boardgame.repository.projection.SituationBoardGameProjection;
 import com.boardgo.domain.boardgame.repository.response.BoardGameByMeetingIdResponse;
 import com.boardgo.domain.boardgame.repository.response.BoardGameSearchResponse;
 import java.util.List;
@@ -13,5 +13,5 @@ public interface BoardGameDslRepository {
 
     List<BoardGameByMeetingIdResponse> findMeetingDetailByMeetingId(Long meetingId);
 
-    List<SituationBoardGameDto> findByMaxPeopleBetween(int maxPeople);
+    List<SituationBoardGameProjection> findByMaxPeopleBetween(int maxPeople);
 }
