+++ src/main/java/com/boardgo/domain/boardgame/repository/projection/CumulativePopularityCountProjection.java
@@ -1,0 +1,3 @@
+package com.boardgo.domain.boardgame.repository.projection;
+
+public record CumulativePopularityCountProjection(Long boardGameId, Long cumulativeCount) {}
