+++ src/main/java/com/boardgo/domain/boardgame/repository/projection/SituationBoardGameProjection.java
@@ -1,9 +1,9 @@
-package com.boardgo.domain.boardgame.repository.dto;
+package com.boardgo.domain.boardgame.repository.projection;
 
 import com.querydsl.core.annotations.QueryProjection;
 
-public record SituationBoardGameDto(
+public record SituationBoardGameProjection(
         String title, String thumbnail, int minPlaytime, int maxPlaytime, String genre) {
     @QueryProjection
-    public SituationBoardGameDto {}
+    public SituationBoardGameProjection {}
 }
